+++ server/routes/product_routes.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Request
 from fastapi.exceptions import HTTPException
 from fastapi.responses import JSONResponse
 
@@ -23,7 +23,7 @@ async def get_products():
                 200, {"error": "An error occurred while getting the content"}
             )
     except HTTPException as error:
-        return {"Error": error.detail}, error.status_code
+        raise HTTPException(status_code=error.status_code, detail=error.detail)
 
 
 @router.get("/{id}", response_model=Product)
@@ -38,14 +38,14 @@ async def get_product_by_id(id: str):
                 200, {"error": "An error occurred while getting the content"}
             )
     except HTTPException as error:
-        return {"Error": error.detail}, error.status_code
+        raise HTTPException(status_code=error.status_code, detail=error.detail)
 
 
 @router.post("/", response_model=dict)
-async def create_product(product: ProductMutation):
-    try:
-        verify_token_admin()
+async def create_product(product: ProductMutation, request: Request):
+    await verify_token_admin(request)
 
+    try:
         image_url, presigned_url = generate_presigned_url("products", product.image)
         product.image = image_url
 
@@ -61,13 +61,13 @@ async def create_product(product: ProductMutation):
                 500, {"error": "An error occurred while creating the content"}
             )
     except HTTPException as error:
-        return {"Error": error.detail}, error.status_code
+        raise HTTPException(status_code=error.status_code, detail=error.detail)
 
 
 @router.put("/{id}", response_model=dict)
-async def update_product(
-    id: str, product: ProductMutation, _=Depends(verify_token_admin)
-):
+async def update_product(id: str, product: ProductMutation, request: Request):
+    await verify_token_admin(request)
+
     try:
         if product.image != "" and product.image.lower() != "null":
             image_url, presigned_url = generate_presigned_url("products", product.image)
@@ -85,11 +85,13 @@ async def update_product(
                 500, {"error": "An error occurred while updating the content"}
             )
     except HTTPException as error:
-        return {"Error": error.detail}, error.status_code
+        raise HTTPException(status_code=error.status_code, detail=error.detail)
 
 
 @router.delete("/{id}", response_model=dict)
-async def delete_product(id: str, _=Depends(verify_token_admin)):
+async def delete_product(id: str, request: Request):
+    await verify_token_admin(request)
+
     try:
         response = await database.delete_product(id)
 
@@ -103,4 +105,4 @@ async def delete_product(id: str, _=Depends(verify_token_admin)):
                 500, {"error": "An error occurred while deleting the content"}
             )
     except HTTPException as error:
-        return {"Error": error.detail}, error.status_code
+        raise HTTPException(status_code=error.status_code, detail=error.detail)
